loplugin:redundantfcast
...after ca357609fe63f45b4ab364618e29e8e32736b8cd "loplugin:redundantfcast
improve to find redundant double casts".

The original code had deliberately used a cast to double as g_variant_get_double
is documented to return gdouble, so it wanted to make sure that the correct type
(rather than e.g. float or long double) is used here to pick the intended <<=
overload.  But at least <https://docs.gtk.org/glib/types.html#gdouble> documents
that gdouble is guaranteed to be: "Equivalent to the standard C double type."
So dropping the cast should indeed be safe.

Reviewed-on: https://gerrit.collaboraoffice.com/c/online/+/1011

+++ engine/configmgr/source/dconf.cxx
@@ -302,7 +302,7 @@ bool getDouble(
             "bad key " << key << " does not match double property");
         return false;
     }
-    *value <<= double(g_variant_get_double(variant.get()));
+    *value <<= g_variant_get_double(variant.get());
     return true;
 }
